refactor: warikanbot.pyをController/Conversation/UI Builderに分割
- line_ui.py: Flex Message・QuickReply構築関数を抽出（純粋関数群）
- conversation.py: OpenAI API操作・ツール呼び出しループを抽出
- warikanbot.py: Controller（WebhookHandler）+ Tool（ディスパッチ）に縮小
- Conversation→Tool依存をtool_executor: Callableで断ち切りDIP達成
- テストで@patchが不要になりcallable mockで直接注入する形に改善
